Split out benchmarks into its own file.    - This enables easier benchmarking across branches without having to do funny merges and cherrypicking Added a runtime test    - because I haven't actually given much thought to the invariant properties of Reshape      hence this is the best foot forwards.    - Ideally we'd want to actually have property based testing on these
